refactor(pii): split validators.rs into per-validator files
Mirror the recognizer/pattern/ layout: one file per validator under
recognizer/validator/ (noop, luhn, iban, us_ssn, ip, and). Co-locate
each private helper with its sole consumer (luhn_passes→luhn.rs,
mod97→iban.rs, collect_digits→us_ssn.rs) and move per-validator unit
tests next to their impls. Public API unchanged — re-exports through
recognizer/mod.rs preserve every external import path.
