+++ crates/pii/src/recognizer/validator/mod.rs
@@ -1,0 +1,15 @@
+//! Built-in validators and the [`AndValidator`] combinator.
+
+mod and;
+mod iban;
+mod ip;
+mod luhn;
+mod noop;
+mod us_ssn;
+
+pub use and::AndValidator;
+pub use iban::IbanValidator;
+pub use ip::IpAddressValidator;
+pub use luhn::LuhnValidator;
+pub use noop::NoopValidator;
+pub use us_ssn::UsSsnValidator;
