+++ crates/pii/src/recognizer/validator/noop.rs
@@ -1,0 +1,13 @@
+//! Default validator that abstains on every candidate.
+
+use crate::recognizer::{ValidationOutcome, Validator};
+
+/// Default validator that abstains on every input.
+#[derive(Debug, Default, Clone, Copy)]
+pub struct NoopValidator;
+
+impl Validator for NoopValidator {
+    fn validate(&self, _candidate: &str) -> ValidationOutcome {
+        ValidationOutcome::Unknown
+    }
+}
